+++ dmr/security/jwt/cookie.py
@@ -56,7 +56,7 @@ class _BaseCookieJWTAuth(ResponseSpecProvider):
         """Declare extra responses for cookie auth + CSRF checks."""
         return [
             *self._add_new_response(
-                unauth_response_spec(controller_cls),
+                unauth_response_spec(controller_cls, metadata),
                 existing_responses,
             ),
             *self._add_new_response(